added minified directory
added a minified directory for better performance with bandwidth
constraints, and added a file path finder in cheap-shots.js for dynamic
pathnames. This will enable users to call the file from anywhere. Note
that 'min' and 'src' directories' internals must still be left intact,
and if 'cheap-shots.js' is called more than once in the document only
the first occurence will be used for PATH variable.